chore: 整理队列控制器与可靠性测试行数
